Fix some 64-to-32 implicit conversions erroring under Ruby 2.4
Prior to Ruby 2.4, Ruby has let these implicit conversions be warnings,
but now they've been changed to full errors. Make the conversions
explicit as they aren't going to be losing any significant information
in actual use.

This fixes Rice to work properly under Ruby 2.4.

+++ rice/Hash.ipp
@@ -292,13 +292,13 @@ begin() const
 inline Rice::Hash::iterator Rice::Hash::
 end()
 {
-  return iterator(*this, size());
+  return iterator(*this, (int)size());
 }
 
 inline Rice::Hash::const_iterator Rice::Hash::
 end() const
 {
-  return const_iterator(*this, size());
+  return const_iterator(*this, (int)size());
 }
 
 inline bool Rice::
